requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
cryptography==41.0.4
sqlalchemy-utils==0.41.1
marshmallow==3.20.1
flask-wtf==1.1.1
wtforms==3.0.1